feat: split editor into frontend and backend components
The backend component runs on the server and the frontend component on the client.
This improves the responsiveness.

The main reason for this refactoring was that the server doesn't know the bounds of
the layouted elements and had to retrieve them from the client for computing properties
such as the code completion menu position.